requirements.txt
django
django-import-export
django-admin-rangefilter